criação do arquivo 02 da lista baby + alteração no arquivo baby1

+++ Lista Baby/baby1.py
@@ -1,6 +1,6 @@
 # Escreva um programa que leia um número e o imprima caso ele seja maior que 10.
 
-numero = int(input("Digite um número: "))
+numero = float(input("Digite um número: "))
 
 if numero < 10:
     print(numero)
